fix: rename pandaspyarrowtransformer.py to use snake_case
Align pandas transformer filename with the convention used by all other
transformer files (e.g. duckdb_pyarrow_transformer.py,
polars_pyarrow_transformer.py).

Closes #261
